chore: moving around some files
Per feedback, rearranging files a bit so as to minimize the number of files called index.jsx

Feedback: https://github.com/Twin-Cities-Mutual-Aid/twin-cities-aid-distribution-locations/pull/302#discussion_r624758043
